Create CountOnesInBit.java
Getting the length of the binary number by performing a right shift on the number with one position till the number becomes zero. Then looping over this length to check each bit by creating a bitmask formed by performing a left shift on every position and performing 'bitwise and' between bitmask and number, if it doesn't give zero means there is 1 present at that position.