Create toggleBit.java
Creating a bit mask by performing a left shift on binary1 by the position that needs to be toggled. Then check if the 'bitwise and' operator between the number and bitmask gives zero if yes, then perform 'bitwise or' between the number and bitmask; else take the negation of bitmask and perform 'bitwise and' between the number and bitmask.