Create selectionSort.java
Selection Sort - Initially, the element at the zeroth index is assumed the smallest, then it is compared with the rest elements. If found smaller element than it, that would be replaced as smallest. After looping over all elements, the smallest element will be swapped by the initial index. Time complexity: O(n^2).