Create insertionSort.java
Initially, the first element is assumed to be in a sorted array, and the rest in an unsorted array. Elements from the unsorted array are compared to sorted elements, if found greater, it will be placed in the sorted array; else, elements from the sorted array will shift to the right giving space to the coming smaller element in the sorted part. Time complexity: O(n^2).