Create recursion_calculatePower.java
Calculating power using recursion: In this, the recursive function takes base and power as parameters, then checks for base condition, if the base is zero then the function will return zero; if not, then the base number will be multiplied by power till the power becomes one. The power is subtracted by 1 in each recursive call.