Create recursion_factorial.java
Factorial using recursion: the recursive function takes an integer input. If the integer parameter is zero or one, the function will return 1. The current integer will be multiplied by the number one less than it. The recursive function will be called having a parameter as the number subtracted by one.