Create recursion_strictlyIncreasingSeries.java
Checking if the given series is strictly increasing using recursion: the recursive function takes the size of the array as one parameter and an integer array as the other parameter. If the size of the array is zero, the function will return true(base case). It will check if the second last element of the array is greater or equal to the last element, if yes then the function will simply return false. The recursive function will call itself with reducing the parameter of size by 1 and the same array.