Create recursion_stackHeightLogn
Calculating the power of a number while keeping the height of the stack as log n, so that the memory allocated at a particular time is less as in general, for calculating power, the stack height is n(value of power). 
In the recursion function, base value and power value are taken as input. If base is zero, then the function will return zero and if the power is one, the function will return 1(base case). For even power, power is divided into two halves and multiplied; else power is divided into two halves and along with their multiplication, base value is also multiplied.