Create recursion_occurrenceOfElement.java
Given a character, the first and the last occurrence of that character in the given string has to be found. Simply, we'll traverse through the string matching each element of the array to a given character, the first it matches, the value of the public variable first will be -1, at that time the index value will be saved in the first occurrence value and last time when the value will match, it will be saved in last occurrence variable.