Create recursion_moveElementInString.java
In this problem, the given element 'x' has to be shifted at the end of the new string. Here two strings are there, the original string whose elements have to be shifted, and the new empty string. In the recursive function, if the character is not 'x' then the character from the original string will be concatenated to the new string. Once we'll reach the length of the original string, we'll hit the base case where there is a loop that runs the number of times the new string is less than the old string. This many times the loop will iterate, adding 'x' to the new string in each iteration.