Create recursion_removeDuplicates.java
Removing duplicate elements from the string. 
Method 1, simply matching if the elements are present in the new string; if not, concatenating the elements in the new string, and if not, just passing on by calling the recursive function by incrementing the index and finally reaching the base case when the index value equals the length of the original string and then printing the new string.
Method 2, using the boolean array having a length of 26 since there are 26 alphabets, each place represents the boolean value for each alphabet. The alphabet subtracted from 'a' gives the index number. (For eg. a - a = 0: 0th index represents the boolean value for a. Similarly, b-a = 1: 1st index represents the boolean value for b ). By default, the boolean array has false values. We'll check if the current of the original string subtracted from 'a' gives true, which means that element of the original string is already present in the new string. If not, then we'll set the value of that index as true and concatenate that element in the new string.  Finally, we'll reach the base case when the index value equals the length of the original string and then printing the new string.