Create backtracking_permutation.java
Printing permutations of the given word. In this basically recursion and backtracking both have been used. 
The recursive function takes two arguments, an input string, and an empty string, to save its permutations. 
Inside the function, we'll run a for loop till the length of the inputted string. Then taking the current character as the character on that index, the rest string will be separated using the substring function. This substring will be not passed to the recursive function, along with the concatenation of the current character into the permutation string. Now, this function will call itself, making substrings till the length of the substrings(originally the inputted string) becomes zero, from there it will backtrack and form permutations and print.