Create recursion_mazeTotalPath.java
Counting the total path to reach the endpoint of the maze. A maze can be taken as a 2-D matrix having 'm' rows and 'n' columns. We have to find the count in which we can traverse from (0,0) to (m-1,n-1) cell. The condition is that from one cell we can only move rightwards or downwards.
For this, we have a recursive function that has four parameters, no. of rows, no. of cols, rows index starting from 0, and cols index starting from 0.  So, first lets give the choice to move downwards, so we'll call the recursive function, increasing the row's index by 1, and saving the end value of this function in a variable(here int downPaths). When the cols index be equal to the length of the cols, there is no more possibility to go in that direction, so it will return 0. Then it will try to move in another direction, so for that we'll call the recursive function for rightward direction, by increasing the cols's index by 1, and saving the end value of this function in a variable(here int rightPaths). 
When the length of both the rows index and cols index equals the rows-1 and cols-1 respectively, the function will return 1 to the last recursive function call.  Then these values of recursive cols are added and returned from the functions as the total count of the paths.

Basically, it will move in a particular direction(either down or right) till it traverses the whole length of that column or row. Then when it reached the length, it changes direction, if both of the paths(both directions) are already traversed then it would backtrack and search for the new path to reach the base case.