Create backtracking_nQueenProblem.java
n-queen problem
The problem: There is an 'n' x 'n' chess board, where we have to place the queens so that no queen can check(take) another queen. Queen walks straight vertically, horizontally, and diagonally. We need to print all such possible queens placement for the given board dimension. If n<4, only one queen can be placed at any position. 

Approach: Using backtracking-recursion.
We take the dimension of the board as input, the chess board will be in the form of a 2-D character array and list of the list of strings for printing the placement of queens in different board arrangments. 
The output basically is the format List<List<String>>, the inner list of string basically is the list of rows, each row represents a string, having queen position as 'Q' and null positions as '.' . So in the inner list, there are 'n' strings that make up the placement of the queen in one board arrangement. Then the outer list is basically the list of all such boards(represented by the inner list). 

There are three functions here:
1) nQueen - the recursive function for placements of the queen.
Parameters: allBoards(ArrayList), board(2-D character array), number of columns(cols) initialized from zero. 
Looping over all rows, for each position checking if the queen is safe(using isSafe function), if it is safe, then that index of the board is given 'Q', basically, a queen can be placed there. Then again calling the recursive function having the same parameters except the column index is incremented by one. If found not safe then, the isSafe function will return false and that index will be given '.', basically queen can't be placed there and from there then it will again start, means again the recursion function will be called incrementing the column value, and check isSafe.
When cols equal the length of the board, then base condition hits, for that board, for that row, we have a result wherever the queen can be placed, so in the base condition, we have a function saveBoards where the placement will be saved.

2) saveBoard -  to save and print all placements of queens on all possible boards using ArrayList.
parameters: List<List<String>> allBoards, board 
for String row, we first create an empty string. Then for one board, we create a list of strings.
Nested Looping over the length of the 2D board, for each row, if the position has 'Q' then appending 'Q' in the row String or else appending '.'. This will happen for all the rows and in this process each row is getting appended in the newBoard and when the row index reaches the length of the board, then we'll get all rows in the newBoard, and then lastly, all this newBoards will be added in the allBoards ArrayList, giving us the final solution.

3) isSafe - boolean function to check if the queen can be placed or not.
parameters: int row, int col, board.
For given index, it will check in all direction that is vertically, horizontally, left upper and lower diagonals and right upper and lower diagonals. Finally, will return true or false accordingly.