Update and rename sudokoSover.java to sudokoSolver.java
Sudoku Solver
A sudoku here is a 9 by 9 matrix having sum given digits between 1-9(inclusive) and there are missing places that has to be filled from number between 1 to 9(inclusive) with following three conditions:
..every digit should be unique in the horizontal row.
..every digit should be unique in the vertical column.
..every digit should be unique in 3 by 3 grid.
To solve this problem, we have first made a Boolean type helper function, in which, a 2-D character board is taken as one of its parameters, the other two being row and col initiated with zero. We have to traverse through the whole 2-D matrix, so base condition is when the row length would equate to the board length, the function will return true. Initially we have we have written a chunk of code for moving in the matrix by assigning new row and column number. Then for if the current position is not null then, it must be a digit so function can just pass on. If it's not a digit then we'll loop over 1 to 9, trying to keep the number at that position, that we'll check from isSafe function. If the returns true, then that number would be placed and then again the helper function will be called for next position, if the helper function returns false, then the function backtracks and puts the dot at the previous position, considering it a wrong choice of number for that position, then for loop increments and another number is placed at that position. When all rows would be traversed, the base case would hit.
The isSafe function basically checks for aforementioned three cases, for checking horizontally and vertically, it simply loops till board length and compare with the given number. For checking the grid, we take out the first index of each grid by dividing the row and col number by 3, getting the quotient and multiplying it again with three, it gives the exact index number of every new grid. Then we apply nested loop for that obtained index till obtained index plus three index and compares with the number, if it matches then function returns false, otherwise true.
Note: here while comparing the number, we've first added the integer to character zero and then typecasted it to character for character array of board. For example if int 1 has to be converted character 1, directly typecasting it will give character ascii value, then is 49. So we first add it to character zero, 0 in ascii is 48, so 48 plus 1 given 49, then typecasting it to character will give character 1.