Update mergeSort.java
Merge sort is a sorting technique in which divide the given array into single elements and then conquer those elements in the mediator array and then copy the final sorted array elements into the original array. 
Here, first we made a function MergeSort, where we divided the array from the middle at every step. Then when we reach the base case that the starting index is equal or greater to the ending index,  this means that we have reached to a stage where we have divided the array that till the single element. Now since the single element array is a sorted array in itself. So from here we trace back, again apply recursive MergeSort function, for dividing the other half of the array, as reach to the base case again. now we have two arrays(for the first case), of single element. We now apply conquer function.
In conquer function, for the first array(first part of divided array from starting index to the middle index), we have a indexing variable idx1, and for other part of the array from middle+1 to ending index, we have indexing element as idx2. We conquer(compare and combine) two arrays(basically two parts of same array : two parts formed from middle). So for the first time conquer function would be called, it has starting index as 0, mid as 0 and ending index as 1. So in while condition, both the condition satisfies, i.e., idx1 is less than or equal to mid and idx2 is less than or equal to ending index, so first while loop will run and first two elements would be compared, in the given case idx2 element is smaller so, merged array would have its 0th element as the second element of the array and then now idx2 will be incremented so first while loop condition won't get satisfied as idx2(2) becomes greater than the ending index(1). So, now the second while loop will run, and next element in merged array would be the first element form the original array. Now we have a for loop,  this result will be copied in the original array.