Update quickSort.java
Quick sort
worst time complexity - 0(n^2)
average time complexity - 0(nlogn)
In this, we basically choose a pivot element and then part array around it. The pivot element can be chosen randomly or the median element or the lowest element or the highest element. 
Here we chose, pivot as the highest element. So in this first we have called the void recursive QuickSort function, in which till the low index number is less than high index number, the we'll going to have the recursive. Basically till we keep on partitioning that the there will be only one single element would be left in the array. 
We take out the index of the pivot element by calling the partition function. When partition function returns the index of the pivot element, we now know the fixed place of the pivot element in the final sorted array. So we call the recursive function for the elements less index of pivot element and next for the elements more than the index of the pivot element. In this way we keep on getting the right index number of pivot element(last index element of the array, since we consider that as our pivot here).
The partition function takes three parameters: the array, low and high(index number). Then we declare the last index element as the pivot. Then we create an index i as low-1 which will keep on making space and placing the element in the array which are less than the pivot element. So we loop from the j=low to the high(excluding high since that is the pivot element). Then if the current element is less than the pivot then i will increment. Swapping of the element with ith element would be done using temporary variable. When this process will be done. Lastly we would place the pivot element. So we'll increment i one more time, using swapping technique, place the pivot element at that place and return this index number to the function. Similarly, we'll do for all the elements and get their index number and get the final sorted array.